Remove failed/sucess status msgs...
They would probably still have worked with the send_message, but due to the fact that edit_message is implemented differently in the telegram bot api we're using, the edit_message was returning weird results and was therefore always saying "failed...".
I removed the checks for the send_message simply to unify them... a bit weird if it sometimes has status msgs and sometimes not.